Query performance improvements
Uses select_related() and prefetch_related() for workers and events
respectively to reduce the amount of extra queries that are undesired
for the usage patterns in SPBM.

select_related() is enough for workers, for most cases, with society
information.
prefetch_related() is necessary for events, as we tend to get the
workers details, and those are associated via the shifts.